优化用户和单据模块

+++ src/main/java/com/jsh/erp/datasource/mappers/DepotHeadMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface DepotHeadMapper {
-    int countByExample(DepotHeadExample example);
+    long countByExample(DepotHeadExample example);
 
     int deleteByExample(DepotHeadExample example);
